+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* Add half_width flag for higlight boxes (PR #520)
 
 ## 9.22.0